+++ core-notifications-fe/src/core-notifications/resources/notifications/UserList.tsx
@@ -4,9 +4,7 @@ import CircleOutlinedIcon from '@mui/icons-material/CircleOutlined';
 import RadioButtonCheckedOutlinedIcon from '@mui/icons-material/RadioButtonCheckedOutlined';
 
 import {
-  Alert,
   Box,
-  Button,
   Link,
   ListItem,
   ListItemAvatar,
@@ -14,9 +12,9 @@ import {
   List as MuiList,
 } from '@mui/material'
 import {
+  Button,
   InfiniteList,
   SingleFieldList,
-  UpdateButton,
   useDataProvider,
   useGetIdentity,
   useListContext,
@@ -31,10 +29,9 @@ const NotificationSubscriber = () => {
 
   const enabled = !!Object.assign({}, dataProvider)?.subscribe
   useSubscribeToRecordList((event) => {
-    console.log('NotificationSubscriber', event)
-
     switch (event.type) {
-      case "created": {
+      case "created":
+      case "updated": {
         refetch();
         break;
       }
@@ -51,7 +48,6 @@ export const UserList = () => {
     <InfiniteList
       resource="Notification"
       sort={{ field: 'type', order: 'DESC' }}
-      // TODO: Enable for production
       filter={{ 
         to_id: [to.id],
         read: false,
@@ -59,8 +55,6 @@ export const UserList = () => {
       perPage={10}
       exporter={false}
       actions={false}
-      // filters={false}
-      // filters={userFilters}
     >
       <SingleFieldList component={MuiList}>
         <NotificationItem />
@@ -79,16 +73,12 @@ const NotificationItem = () => {
     <ListItem
       alignItems="flex-start"
       sx={{ display: 'flex', justifyContent: 'space-between', width: '100%' }}
-      // secondaryAction={<ActionButtons />}
     >
-      {/* <ListItemButton> */}
       <ListItemAvatar>
         {notification.type.includes('Task') ? <TaskAltOutlinedIcon color="warning" /> : (notification.read_at ? <RadioButtonCheckedOutlinedIcon /> : <CircleOutlinedIcon />)}
       </ListItemAvatar>
       <ListItemText primary={notification.title} secondary={notification.description} />
-
       <ActionButtons />
-      {/* </ListItemButton> */}
     </ListItem>
   )
 }
@@ -96,31 +86,41 @@ const NotificationItem = () => {
 const ActionButtons = () => {
   const notification = useRecordContext()
   const { data: currentUser } = useGetIdentity()
+
   const [update] = useUpdate()
 
-  const onUpdateSuccess = () => {
-    console.log('updateSuccess')
-  }
 
-  const markAsReadArgs = notification.type.includes('Global') ? { id: currentUser?.id} : null
+  const markAsRead = (e: any) => {
+    e.preventDefault() // necessary to prevent redirects on update (default save behavior)
 
-  const markAsRead = () => {
     update(
       'Notification',
       {
         id: notification.id,
         data: {
           instance_method: 'mark_as_read!',
-          instance_method_args: markAsReadArgs
+          instance_method_args: notification.type.includes('Global') ? { id: currentUser?.id} : null
 
         },
         previousData: notification,
-      },
-      { 
-        mutationMode: 'optimistic',
-        onSuccess: onUpdateSuccess
       }
     )
+  }
+
+  const markAsComplete = (e: any, action_response: string) => {
+    e.preventDefault()
+
+    update(
+      'Notification',
+      {
+        id: notification.id,
+        data: {
+          instance_method: 'mark_as_complete!',
+          instance_method_args: { action_response }
+        },
+        previousData: notification,
+      }
+    ) 
   }
 
 
@@ -129,15 +129,13 @@ const ActionButtons = () => {
       <Box sx={{ display: 'flex', flexWrap: 'wrap', gap: 2, ml:2}}>
         {
           notification.actions.map((action: {label: string; response: string }, i: number) => (
-            <UpdateButton
+            <Button
               key={i}
               startIcon={undefined}
               color={i == 0 ? "primary" : "secondary"}
-              variant="outlined"
+              variant={i == 0 ? "contained": "outlined" }
+              onClick={(e) => markAsComplete(e, action.response)}
               label={action.label}
-              data={{ instance_method: 'mark_as_complete!', instance_method_args: { action_response: action.response } }}
-              mutationMode='optimistic'
-              mutationOptions={{ onSuccess: onUpdateSuccess }}
             />
           ))
         }
@@ -146,16 +144,5 @@ const ActionButtons = () => {
 
   if (notification.link) return <Link onClick={markAsRead} href={notification.link.url}>{notification.link.label || 'Read More...'}</Link>
 
-  return (
-    <UpdateButton 
-      startIcon={undefined} 
-      sx={{whiteSpace: 'nowrap'}} 
-      label="Mark as Read" 
-      data={{ instance_method: 'mark_as_read!', instance_method_args: markAsReadArgs }} 
-      mutationMode='optimistic'
-      mutationOptions={{ onSuccess: onUpdateSuccess }}
-    />
-  )
+  return <Button startIcon={undefined} sx={{whiteSpace: 'nowrap'}} onClick={markAsRead} label='Mark as Read' />
 }
-
-// const userFilters = [<SearchInput key="q" source="q" alwaysOn />]